--- a/requirements.txt
+++ b/requirements.txt
@@ -5,6 +5,5 @@ pylint
 vcrpy
 contextlib2
 mock
-jsonpickle
 joblib
 tqdm