--- a/src/README.rst
+++ b/src/README.rst
@@ -18,7 +18,7 @@ Change Log
 
 Unreleased
 ----------
-- Updated 'application upload' native imagestore application upload to parallelize uploads, greatly reducing overall upload time for applications with many smaller files. Also improved progress log shown by flag '--show-progress'. (#234)
+- Updating 'application upload' native imagestore application upload to parallelize uploads, greatly reducing overall upload time for applications with many smaller files. Also improving progress log shown by flag '--show-progress'. (#234)
 
 9.0.0
 ----------