@samedder 2020-11-05
> @samedder where can we check to make sure these 2 new libraries are allowed to be used?

There's two options, you can merge it and see if internally it generates a bug for insecure package use. That will require a PR to follow up. Or, you can do what I did and try to find another repo already using them and verify they don't have an open issue.

I looked manually now though and it seems the packages are O.K. but I didn't check the specific version.